requirements.txt
# requirements.txt
# Framework web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Google Maps API
googlemaps==4.10.0

# Base de données et cache
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# Géospatial
geopandas==0.14.1
shapely==2.0.2

# Analyse de données
numpy==1.26.2
pandas==2.1.3

# Utilitaires
python-dotenv==1.0.0
requests==2.31.0

# Tests
pytest==7.4.3
httpx==0.25.2